+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.7.6 (02/2023)
 1. Focal support added
 1. New focal-compatible binaries and libs included
+1. Upstart migrated to systemd
 
 # 0.7.5 (04/2022)
 1. UI re-designed a bit - new main screen in landscape mode